refactor: don't revert deployment on non-existant configs (#15)
At this point, the contracts that need to be deployed do not require
chain specific configurations. The ones provided for local test nodes
are used in tests that don't represent the production deployment.

Hence, there's little point in reverting the deployment if there's no
config for a specific chain.

This commit also moves the `deployer` assignment to the beginning of the
constructor, ensuring that it's set when configs are created.